requirements.txt
streamlit
google-generativeai
langchain
python-dotenv
PyPDF2
faiss-cpu
langchain_google_genai